# api/docs/openapi.yaml
openapi: 3.0.3
paths:
...
  /organizations/{organization_uuid}/integrations:
    get:
      operationId: IntegrationsController_findAll
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/tools:
    get:
      operationId: IntegrationsController_getEnabledTools
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/{integration_uuid}:
    get:
      operationId: IntegrationsController_findOne
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
    patch:
      operationId: IntegrationsController_update
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIntegrationDto"
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
    delete:
      operationId: IntegrationsController_delete
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/{integration_uuid}/test:
    post:
      operationId: IntegrationsController_testConnection
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/{integration_uuid}/actions:
    get:
      operationId: IntegrationsController_getActions
      parameters:
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/{integration_uuid}/actions/{action_uuid}:
    patch:
      operationId: IntegrationsController_toggleAction
      parameters:
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
        - name: action_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleIntegrationActionDto"
      responses:
        "200":
          description: ""
      tags:
        - Integrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/database-integrations:
    post:
      operationId: DatabaseIntegrationsController_createDatabaseIntegration
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatabaseIntegrationDto"
      responses:
        "201":
          description: ""
      tags:
        - DatabaseIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/database-integrations/test:
    post:
      operationId: DatabaseIntegrationsController_testDraftConnection
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestDatabaseConnectionDto"
      responses:
        "201":
          description: ""
      tags:
        - DatabaseIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/database-integrations/{integration_uuid}:
    get:
      operationId: DatabaseIntegrationsController_getDetails
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - DatabaseIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/database-integrations/{integration_uuid}/sync:
    post:
      operationId: DatabaseIntegrationsController_syncSchema
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - DatabaseIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/database-integrations/{integration_uuid}/test:
    post:
      operationId: DatabaseIntegrationsController_testSavedConnection
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - DatabaseIntegrations
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/integrations/mcp/test-connection:
    post:
      operationId: McpIntegrationsController_testConnectionDraft
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestMcpConnectionDto"
      responses:
        "201":
          description: ""
      tags:
        - McpIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/mcp:
    post:
      operationId: McpIntegrationsController_create
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMcpIntegrationDto"
      responses:
        "201":
          description: ""
      tags:
        - McpIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/mcp/{integration_uuid}:
    get:
      operationId: McpIntegrationsController_getDetails
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - McpIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/mcp/{integration_uuid}/sync-tools:
    post:
      operationId: McpIntegrationsController_syncTools
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - McpIntegrations
      security:
        - bearer: []
  /organizations/{organization_uuid}/integrations/mcp/{integration_uuid}/test:
    post:
      operationId: McpIntegrationsController_testConnection
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: integration_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - McpIntegrations
      security:
        - bearer: []
...
  /organizations/{organization_uuid}/documents:
    post:
      operationId: DocumentsController_upload
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - Documents
      security:
        - bearer: []
      summary: Upload a document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
    get:
      operationId: DocumentsController_findAll
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Documents
      security:
        - bearer: []
  /organizations/{organization_uuid}/documents/{document_uuid}/widget-content:
    get:
      operationId: DocumentsController_getWidgetContent
      parameters:
        - name: organization_uuid
          required: true
          in: path
          schema:
            type: string
        - name: document_uuid
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Documents
      security:
        - bearer: []
...
info:
  title: Cortex API
  description: >-
    REST API for Cortex — organizations, conversations, agents, integrations (Composio, MCP, OpenAPI, databases),
    documents, roles, and members. Authenticate via JWT Bearer token from POST /auth/email/login.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3001
    description: Local
  - url: "{{baseUrl}}"
    description: Environment variable
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/email/login
  schemas:
...
    UpdateIntegrationDto:
      type: object
      properties: {}
    ToggleIntegrationActionDto:
      type: object
      properties: {}
    CreateDatabaseIntegrationDto:
      type: object
      properties: {}
    TestDatabaseConnectionDto:
      type: object
      properties: {}
...
    TestMcpConnectionDto:
      type: object
      properties: {}
    CreateMcpIntegrationDto:
      type: object
      properties: {}
...
security:
  - bearer: []